kfctl - Makefile and Dockerfile enhancements (#3189)
* bootstrap/Makefile: export kfctl binary in build-kfctl-container target

This commit extends the build-kfctl-container target so that
it also exports the kfctl binary from the kfctl container.
This is useful because it provides a reproducible way to
generate the kfctl binary.

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>

* bootstrap/Dockerfile: optimize Docker caching

This commit changes the Dockerfile so that go.mod and go.sum
are copied in before the code, so that dependencies are cached
by Docker.

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>