centraldashboard: redirect non-iframed apps (#5469)
* centraldashboard: Rename CM in a more general way

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>

* centraldashboard: Add settings API

Add a `/dashboard-settings` API to retrieve the `settings` field of the
Centrldashboard API.

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>

* centraldashboard: Redirect non-iframed apps

If a KF web app imports the Centraldashboard's shared library, it means
that it wants to integrate with the Centraldashboard (obviously). The
Centraldashboard exposes the web apps via an iframe, so that the app
itself is oblivious of this.

For this reason, all URLs present inside the app point directly to the
app's endpoint and, in the case a user opens a link directly, they will
loose the Centraldashboard context.

This commits changes this behavior: whenever an app imports the shared
library and calls the `init` function, the window location is changed to
prepend the `/_/` centraldashboard's prefix. In this way the app will
find itself inside the iframe and the overall KF UI context will not be
lost.

This behavior is optional and configurable either globally or on a
per-app basis:

* Global config: set `DASHBOARD_FORCE_IFRAME` to `true` under the
ConfigMap `settings` field
* Per-app config: set the `disableForceIframe` flag in the `init`
function if you want your app to opt out of iframe redirection

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>